Extrack FireButton to separate file
